Link framework to tensorrt lib

+++ onnxruntime/core/providers/tensorrt/tensorrt_execution_provider.cc
@@ -693,7 +693,7 @@ std::unique_ptr<IndexedSubGraph> TensorrtExecutionProvider::GetSubGraph(SubGraph
         if (it->GetDstArgIndex() < static_cast<int>(it->GetNode().InputDefs().size())) {
           output = (it->GetNode()).InputDefs()[it->GetDstArgIndex()];
         } else {
-          output = (it->GetNode()).ImplicitInputDefs()[it->GetDstArgIndex() - it->GetNode().InputDefs().size()];
+          output = (it->GetNode()).ImplicitInputDefs()[it->GetDstArgIndex() - static_cast<int>(it->GetNode().InputDefs().size())];
         }
         if (node_set.find(node_idx) != node_set.end()) {
           const auto& iter = fused_inputs.find(output);
